requirements.txt
# Core dependencies
requests>=2.25.0
pandas>=1.3.0,<2.0.0
numpy>=1.19.0,<1.25.0
matplotlib>=3.3.0,<4.0.0

# Geospatial and mapping
folium>=0.12.0,<0.15.0
geopandas>=0.9.0,<0.15.0
rasterio>=1.2.0,<1.4.0
shapely>=1.7.0,<3.0.0

# Earth observation alternatives
earthengine-api>=0.1.300
# Removed sentinel-hub - use Google Earth Engine instead

# AI/ML
openai>=0.27.0
python-dotenv>=0.19.0

# Data processing
scipy>=1.7.0,<1.12.0
scikit-learn>=1.0.0,<1.4.0

# Jupyter notebook support
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Optional: Image processing
pillow>=8.0.0,<10.0.0
opencv-python-headless>=4.5.0,<5.0.0

# Web scraping (for historical documents)
beautifulsoup4>=4.9.0
selenium>=4.0.0
